requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
Pillow==10.1.0
pdf2image==1.17.0
aiofiles==23.2.1
numpy>=1.24.0